+++ naturtag/cli.py
@@ -22,7 +22,7 @@ from rich.table import Column, Table
 
 from naturtag.constants import CLI_COMPLETE_DIR
 from naturtag.metadata import KeywordMetadata, MetaMetadata, refresh_tags, tag_images
-from naturtag.settings import Settings, setup
+from naturtag.storage import Settings, setup
 from naturtag.utils import get_valid_image_paths, strip_url
 
 CODE_BLOCK = re.compile(r'```\n\s*(.+?)```\n', re.DOTALL)
@@ -209,15 +209,17 @@ def refresh(recursive, image_paths):
     click.echo(f'{len(metadata_objs)} Images refreshed')
 
 
-@main.command()
-@click.pass_context
-@click.option('-a', '--all', is_flag=True, help='Install all features')
-@click.option('-d', '--db', is_flag=True, help='Initialize taxonomy database')
+@main.group(name='setup')
+def setup_group():
+    """Setup commands"""
+
+
+@setup_group.command()
 @click.option(
-    '-s',
-    '--shell',
-    type=click.Choice(['bash', 'fish']),
-    help='Install shell completion scripts',
+    '-d',
+    '--download',
+    is_flag=True,
+    help='Download taxonomy data if it does not exist locally',
 )
 @click.option(
     '-f',
@@ -225,13 +227,41 @@ def refresh(recursive, image_paths):
     is_flag=True,
     help='Reset database if it already exists',
 )
-def install(ctx, all, db, shell, force):
-    """Install shell completion and other features.
+def db(download, force):
+    """Set up Naturtag's local database.
+
+    Naturtag uses a SQLite database to store observation and taxonomy data. This command can
+    initialize it for the first time, reset it, or download missing data for taxon text search.
 
     \b
-    Shell tab-completion is available for bash and fish shells. To install, run:
+    Example: Full reset and download, with debug logs:
     ```
-    nt install -s [shell name]
+    nt -vv setup db -f -d
+    ```
+    """
+    click.echo('Initializing database...')
+    setup(overwrite=force, download=download)
+
+
+@setup_group.command()
+@click.option(
+    '-s',
+    '--shell',
+    type=click.Choice(['bash', 'fish']),
+    help='Install completion script for a specific shell only',
+)
+def shell(shell):
+    """Install shell tab-completion for naturtag.
+
+    \b
+    Completion is available for bash and fish shells. To install, run:
+    ```
+    nt setup shell
+    ```
+
+    Or for a specific shell only:
+    ```
+    nt setup shell -s [shell name]
     ```
 
     \b
@@ -240,16 +270,7 @@ def install(ctx, all, db, shell, force):
     nt tag -t corm<TAB>
     ```
     """
-    if not any([all, db, shell]):
-        click.echo('Specify at least one feature to install')
-        click.echo(ctx.get_help())
-        ctx.exit()
-
-    if all or shell:
-        install_shell_completion('all' if all else shell)
-    if all or db:
-        click.echo('Initializing database...')
-        setup(overwrite=force)
+    install_shell_completion(shell or 'all')
 
 
 def enable_logging(level: str = 'INFO', external_level: str = 'WARNING'):
@@ -393,5 +414,5 @@ def _install_bash_completion():
     print(f'source {completion_dir}/*.bash\n')
 
 
-for cmd in [tag, refresh, install]:
+for cmd in [tag, refresh, db, shell]:
     cmd.help = colorize_help_text(cmd.help)